Move global search and peakfinding into search/
