+++ src/subhkl/io/parser.py
@@ -1324,7 +1324,7 @@ def zone_axis_search(
     import jax.numpy as jnp
     from subhkl.config import beamlines, reduction_settings
     from subhkl.optimization import FindUB
-    from subhkl.prior import HoughDavenportPrior, ImageBasedObjective
+    from subhkl.search.prior import HoughDavenportPrior, ImageBasedObjective
     from subhkl.utils import generate_reflections
 
     print(f"Loading data from {merged_h5_filename}...")
