+++ src/subhkl/search/sparse_rbf.py
